Added the EventualConsistencyMiddleware. One issue I see so far is with data integrity when it comes to foreign key constraints

+++ src/GymManagement.Application/Subscriptions/Events/SubscriptionDeletedEventHandler.cs
@@ -12,8 +12,12 @@ public class SubscriptionDeletedEventHandler(ISubscriptionsRepository subscripti
 
     public async Task Handle(SubscriptionDeletedEvent notification, CancellationToken cancellationToken)
     {
-        var subscription = await _subscriptionsRepository.GetByIdAsync(notification.SubscriptionId)
-                           ?? throw new InvalidOperationException();
+        var subscription = await _subscriptionsRepository.GetByIdAsync(notification.SubscriptionId);
+
+        if (subscription is null)
+        {
+            throw new InvalidOperationException();
+        }
         
         await _subscriptionsRepository.RemoveSubscriptionAsync(subscription);
         await _unitOfWork.CommitChangesAsync();
